requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
aiohttp==2.3.10
aiohttp_cors==0.5.3
aiomysql==0.0.15
async-timeout==3.0.0
attrs==18.1.0             # via pytest
chardet==3.0.4
click==6.7
more-itertools==4.1.0     # via pytest
multidict==4.3.1
passlib==1.7.1
pluggy==0.6.0             # via pytest
py==1.5.3                 # via pytestf
pyjwt==1.6.4
pymysql==0.7.11
pytest-aiohttp==0.3.0
pytest==3.6.0            # via pytest-aiohttp
python-dotenv==0.7.1
#pyyaml==3.12
six==1.11.0               # via more-itertools, pytest
sqlalchemy==1.2.7
typing==3.6.4             # via aiohttp-cors
uvloop==0.9.1
yarl==1.2.4
xlrd==1.1.0
aiohttp-swagger==1.0.5
asn1crypto==0.24.0
crypto==1.4.1
aiojobs==0.2.1
#pycrypto==2.6.1
#boto==1.7.32
#cmd :sudo /usr/local/bin/python3 -m pip install -r requirements.txt
